+++ src/main/java/com/utp/sistema/ventas/model/dao/impl/ArticuloDaoImp.java
@@ -66,7 +66,7 @@ public class ArticuloDaoImp extends DataBase implements ArticuloDao {
                 articulo.setIdArticulo(rs.getInt(1));
                 articulo.setIdCategoria(rs.getInt(2));
                 articulo.setDescripcion(rs.getString(3));
-                articulo.setPrecio_venta(rs.getDouble(4));
+                articulo.setPrecio_venta(rs.getFloat(4));
                 articulo.setStock(rs.getInt(5));
                 articulo.setEstado(rs.getByte(6));
             } else {
@@ -100,7 +100,7 @@ public class ArticuloDaoImp extends DataBase implements ArticuloDao {
                 articulo.setIdArticulo(rs.getInt(1));
                 articulo.setIdCategoria(rs.getInt(2));
                 articulo.setDescripcion(rs.getString(3));
-                articulo.setPrecio_venta(rs.getDouble(4));
+                articulo.setPrecio_venta(rs.getFloat(4));
                 articulo.setStock(rs.getInt(5));
                 articulo.setEstado(rs.getInt(6));
